+++ frontend/src/metabase/query_builder/components/view/QuestionTableToggle/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./QuestionTableToggle";
